+++ pma-sync/src/main/java/com/ftn/pmasync/repository/ReportRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ftn.pmasync.model.Report;
 
-public interface ReportRepository extends JpaRepository<Report, Long>{
+public interface ReportRepository extends JpaRepository<Report, Integer>{
+	
+	//Report fineOneById(int id);
 
 }
