+++ pma-sync/src/main/java/com/ftn/pmasync/repository/UserRepository.java
@@ -5,6 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ftn.pmasync.model.User;
 
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
+	
+	//User findOneById(int id);
 
 }
